hotel_service/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
djangorestframework==3.15.2
grpcio==1.70.0
grpcio-tools==1.70.0
idna==3.10
protobuf==5.29.3
psycopg2==2.9.10
pybreaker==1.2.0
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
